Fixed number tags shortValue() being casted to the wrong type

+++ MCStructs-nbt/src/main/java/net/lenni0451/mcstructs/nbt/tags/DoubleTag.java
@@ -51,7 +51,7 @@ public class DoubleTag implements INbtNumber {
 
     @Override
     public short shortValue() {
-        return (byte) (((int) Math.floor(this.value)) & 0xFFFF);
+        return (short) (((int) Math.floor(this.value)) & 0xFFFF);
     }
 
     @Override
